+++ pycobertura/cli.py
@@ -1,19 +1,39 @@
 import click
 
 from pycobertura.cobertura import Cobertura
-from pycobertura.reports import TextReport, TextReportDelta
+from pycobertura.reporters import (
+    HtmlReporter, TextReporter, HtmlReporterDelta, TextReporterDelta
+)
 
 
 pycobertura = click.Group()
 
 
+reporters = {
+    'html': HtmlReporter,
+    'text': TextReporter,
+}
+
+
 @pycobertura.command()
 @click.argument('cobertura_file')
-def show(cobertura_file):
+@click.option(
+    '-f', '--format',
+    default='text',
+    type=click.Choice(list(reporters))
+)
+def show(cobertura_file, format):
     """show the coverage summary"""
     cobertura = Cobertura(cobertura_file)
-    report = TextReport(cobertura)
+    Reporter = reporters[format]
+    report = Reporter(cobertura)
     print(report.generate())
+
+
+delta_reporters = {
+    'text': TextReporterDelta,
+    'html': HtmlReporterDelta,
+}
 
 
 @pycobertura.command()
@@ -22,10 +42,17 @@ def show(cobertura_file):
 @click.option(
     '--color/--no-color', default=None,
     help='Colorize the output. By default, pycobertura emits color codes only '
-         'when standard output is connected to a terminal.')
-def diff(cobertura_file1, cobertura_file2, color):
+         'when standard output is connected to a terminal. This has no effect '
+         'with the HTML output format.')
+@click.option(
+    '-f', '--format',
+    default='text',
+    type=click.Choice(list(delta_reporters))
+)
+def diff(cobertura_file1, cobertura_file2, color, format):
     """compare two coverage files"""
     cobertura1 = Cobertura(cobertura_file1)
     cobertura2 = Cobertura(cobertura_file2)
-    report = TextReportDelta(cobertura1, cobertura2, color=color)
+    Reporter = delta_reporters[format]
+    report = Reporter(cobertura1, cobertura2, color=color)
     print(report.generate())