+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * add `--color` and `--no-color` options to `pycobertura diff`.
+* add option `-f` and `--format` with output of `text` (default) and `html`.
+* renamed *[Rr]eport -> *[Rr]eporter
 
 ## 0.0.2 (2014-11-27)